Move mock database modules to `androidTest` folder
